Farm_Direct_Connect_Full_Stack/backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
python-decouple==3.8
django-cors-headers==4.0.0
django-filter==23.1
celery==5.3.0
redis==4.5.5
channels==4.0.0
channels-redis==4.1.0
razorpay==1.3.0
requests==2.31.0
pillow==9.5.0
django-extensions==3.2.3